python-requirements.txt
beautifulsoup4
requests
selenium
kafka-python
